Change: merge script_calls_mandatory into script_tags_mandatory
